lgp: improve package resolution system by searching in the lgp result dir at first
This change is mandatory when you're building a lot of inter-dependant packages.
The default will be to add '~/dists/$DIST' in the header of the /etc/apt/sources.list

RESULDIR is now exported in the chroot environment (but read-only)

CAUTION! Be sure to have a valid RESULDIR directory and not a symlink because it
will prevent from the auto-mount.