Make lgp tag not require pbuilder base tarballs (closes #66580)
Drop the 'architecture' attribute from setupinfo, and the $arch variable
for tags.  There doesn't seem to be any reason to include a list of
architectures in tag names, and the function to get that list wants to
look at the pbuilder tarballs, which might not exist if we just want to
tag.